Make `hash_combine` support enums automatically
Summary:
Lots of boilerplate to define hash functions for different enums, so we can use them in `hash_combine`.

This change makes it so that `hash_combine` will automatically hash underlying type of enums.

I chose not to specialize enums generically with `std::hash` functor because it feels like we shouldn't leak that. But `hash_combine` we can make this change to.

Changelog: [Internal]

Differential Revision: D53074535