Cleanup srcSet for java and res (#42659)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/42659

Those folders for Java don't exist anymore, I'm removing this as it's unnecessary and will default to only src/main/java.
For resources instead, I'm using `setSrcDirs` as it will replace the default, while `srcDirs()` will add those folders.
We need to replace the default res folder as we need to follow the resource folder structure of BUCK

Changelog:
[Internal] [Changed] - Cleanup srcSet for java and res

Reviewed By: cipolleschi

Differential Revision: D53083677

fbshipit-source-id: 4dc42c700ea5446bbd49c63fc43b58ba316f4944