Fix missing packager resources after applying an OTA update
Summary:
## Changelog:
[Internal]-

There was an heuristic in the asset resolving logic RN code (coming from back in 2016), whereas if JS bundle URL starts from `file://`, then all of the local image paths are expected to be relative to the JS bundle path itself (this is Android-only).

This is not the case in certain scenarios, when the new JS bundle is located in a `file://...` folder, however all of the assets that were served from the local file system and coming from the APK should continue doing so, as there is no images being downloaded with the JS bundle patch update - indeed, such images should be served as remote assets.

The diff makes sure that the local files continue to be served from the correct paths.

Differential Revision: D53407804