Fix compatibility issue with 'cp' command in shell script.
This change addresses a compatibility issue arising from the use of the
macOS-specific '-X' flag in the 'cp' command withing a shell script
called during the 'npm run ios' command.

The '-X' flag prevents the copying of Extended Attributes and resource
forks, which indeed prevents possible issues on macOS. However the flag
is not supported by GNU's 'cp', commonly installed via homebrew on macOS
systems, and usually added to the user's PATH.

This fix basically does a feature test to figure out if the '-X' flag is
available, and only uses it is the test is possitive.

The intended behavior of not copying Extended Attributes and resources
forks is still preserved in the script even if GNU's cp is in the path,
as GNU's cp doesn't copy these by default.

In short, the script will behave correctly regardless of the 'cp'
version installed, but the script will now be more POSIX compatible.

The fix helps prevent unnecessary debugging time as, in my experience,
the source of this error is not always displayed in the output of 'npm
run ios.