Bugfix - FlatList Multicolumn
The keyExtractor for multicolumn FlatList was causing all items in the list to remount. This is because the item received by the keyExtractor in case of multicolumn FlatList is a row wrapper - not the item. I changed the return value of keyExtractor for this case in order to stop FlatList items from remounting.