Unify BUCK and CMake includes for Cxx modules (#43148)
Summary:

This diff adds the `react/renderer/components` prefix to the generated `CMakeLists.txt` file for the Cxx modules.
This will allow to import the generated JSI header also like this
```
#include <MyModule/MyModuleJSI.h>
```
and not just like this
```
#include "MyModuleJSI.h"
```
This will make CMake includes consistent with BUCK includes, and we'll reduce the number of awkward ifdefs in the include section.

Changelog: [General][Changed] - Unify BUCK and CMake includes for Cxx modules.

Differential Revision: D54063743