Add UniquePtrFactory.back, use most recent remote connection in ReactInstanceIntegrationTest
Summary:
Because we use `mockRemoteConnections_.make_unique()` on each `ReactInstanceIntegrationTest::SetUp`, the current implementation of `getRemoteConnection` (which accesses `mockRemoteConnections_[0]`) is only correct after the test is set up *for the first time*, and fails in (e.g.) parameterised tests.

This modifies it to use the last `MockRemoteConnection` created by the factory, for which (for convenience+clarity) we add `back()` to `UniquePtrFactory`.

Changelog: [Internal]

Differential Revision: D54158008