Check whether _bridge and surfacePresenter are not nil to prevent crashes.
Summary:
Add some nil checks to make sure that the objects are there before calling methods on it.
Ususally,in ObjectiveC, this should not be a problem, but the surfacePresenter is an associated object, so I"m not sure that behaves as normal properties for the runtime.

## Changelog:
[Internal] - Add nil check to verify that the app is not crashing

## Facebook:
See S397861 and T180527210 for more information.

Differential Revision: D54246220