Do not invalidate RCTInstance if jsbundle failed to load
Summary:
With this diff we will not be invalidating the `RCTInstance` in case of `jsbundle` loading error.
This change will allow us to to show the RedBox error message. We couldn't do that previously because we weren't able to instantiate the "RedBox" module for the the `RCTTurboModuleRegistry` that was in invalid state together with `RCTInstance`.

Changelog: [iOS][Fixed] - Show RedBox error message if Metro instance is not running.

Differential Revision: D54364278