Modularise scripts directory as Yarn workspace
Summary:
Adds a `package.json` to `scripts/`, making it a Yarn workspace package, `react-native/scripts`. We are doing this so that call sites in Meta's codebase are able to reference script utils without installing the entire monorepo.

Moves unique deps from project root into `scripts/package.json` (note: the vast majority of `devDependencies` are used both inside/out of `scripts/` and could not be encapsulated).

NOTE: Probably not the end state for package management under `scripts/`. We may in future subpackage folders under scripts for further modularisation (e.g. `scripts/build/`, and as with `tools/*` packages currently), but not quite now.

Changelog: [Internal]

Differential Revision: D55062745