decouple RCTBridge+Private from jsinspector-modern
Summary:
Changelog: [Internal]

RCTBridge+Private is transitively bringing in boost in 0.74, which is causing build errors for some libs that were originally depending on it. this is because boost is a special pod that needs separate handling to link correctly.

let's simplify this problem by decoupling this header and moving the inspector specific logic into it's own category.

Differential Revision: D55228474