Fix ReactEventEmitter on bridge-mode (#43641)
Summary:

When on bridge-mode, users are reporting that events are not dispatched well on Android.
This is happening because `ReactEventEmitter` is checking wether the view is on Fabric or Paper,
and dispatching the event properly.

When on Bridge mode, on a Paper legacy view, the `uiManagerType` would be `UIManagerType.FABRIC` (as the RootView is set with Fabric Enabled),
but the `mFabricEventEmitter` will be null. So this results on events being completely ignored.

I've fixed this by checking if the user has enabled Fabric Interop mode, and if so, use the default event emitter.
Fixes #43635

Changelog:
[Android] [Fixed] - Fix ReactEventEmitter on bridge-mode

Differential Revision: D55317418