Fix `caretHidden` prop when text color set
Summary:
Android `setCursorVisible` does not seem to work correctly when text is present, even if set by ForegroundColorSpan. I did some diving to see if I could root cause this within Android, but couldn't find. So this applies a hack (that we also use in FDS) to hide via transparent drawable instead of builtin API.

Changelog:
[Androoid][Fixed] - Fix `caretHidden` prop when text color set

Differential Revision: D55377434