Introduce fragment-based navigation
Summary:
Currently, on Twilight Android every time a new screen is opened a new XOCMainActivity is created. This is not ideal as this activity has a lot of one-time logic that's core to app startup (e.g. setting up touch event listeners, notifications, Bloks, asking for permissions) that is not needed after app startup and definitely not necessary when simply opening a new screen on the app.

This introduced fragment-based navigation which, instead of creating a new activity, creates a new fragment - making use of the current XOCMainActivity. This brings about several benefits:
- Performance: We no longer waste time/resources re-creating things that have already been initialized on app start.
- User experience:
    - We can now persist a bottom tab bar at the bottom of the app (in parity with iOS) which makes it a lot easier for a user to navigate in the app. Previously, it was easy to get lost within screens and the only way to navigate back to the tabs were to go all the way back in the stack. (cc alikherbane)
    - Previously, because we were re-creating the activity, permissions would be checked and requested on every screen open. This is super disruptive as in the worst case if the user declined every time, they would be asked again as soon as they open another screen like PDP.

## Changelog:
[Internal] [Changed] - Add option to skip calling delegate on ReactFragment lifecycle events

Differential Revision: D55533768