Fix measure not using the last reported mark with a given name (#43703)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/43703

Changelog: [internal]

(internal because this API isn't available in OSS yet)

I found a bug in the current implementation of `performance.measure` where the API would use the first `mark` reported under a specific name instead of the last one (found it in the new example in RNTester in D55477746 that re-logs the marks every time we click on a button).

The root cause for this problem is that we were using `insert` from `std::unordered_set` to update the value, but `insert` doesn't modify the value if it's already present.

This fixes the issue by doing a lookup and removing the value prior to inserting it.

Reviewed By: rshest

Differential Revision: D55477743

fbshipit-source-id: e72aa784a936828db64b572988fe0acb2ad78214