Fix kotlinification of RuntimeConfig (#43947)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/43947

In D55788661, we kotlinified RuntimeConfig:

Which meant converting this:

```
public final class RuntimeConfig {
  public long heapSizeMB;
}
```

To
```
public class RuntimeConfig {
  public var heapSizeMB: Long = 0L
}

```

## Problem

HermesExecutor uses RuntimeConfig like this:

```
HermesExecutor(Nullable RuntimeConfig config, boolean enableDebugger, String debuggerName) {
    super(
        config == null
            ? initHybridDefaultConfig(enableDebugger, debuggerName)
            : initHybrid(enableDebugger, debuggerName, config.heapSizeMB));
  }
```

But, heapSizeMB is no longer a "java property" on RuntimeConfig. I think a property exists for heapSizeMB, but it's private on RuntimeConfig. Instead, java is expected to use the getter and setters, which are public. So, compiling HermesExecutor fails, because HermesExecutor is trying to access a private thing on RuntimeConfig.

This diff just migrates HermesExecutor (java) to use the getter for heapSizeMB, which fixes the error.

Changelog: [Internal]

Reviewed By: fabriziocucci, zeyap

Differential Revision: D55821869