RenderScript Impl of Box Shadow (#43988)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/43988

The supported path for blur composition effects on API 31+ is RenderNode, but that is pretty new.

This adds a RendderScript path, deprecated for newer versions of Android.

We still need API 26+ for clipOutPath (we need to clip original BG out of shadow, e.g. for transparent BG), but this should be supported in many more places.

It is not perfect. RenderScript's built-in blur implementation has a maximum physical pixel radius of 25px. On an xxhdpi device, this means shadows stop getting larger after ~8 CSS pixels. I also think this path can end up with  blurry bitmaps if scale transform enlarged, unlike RenderNode path.

I think this is probably an acceptable degradation path for older devices, and probably not work home-rolling more complex RenderScript shaders. We should potentially add an OS API level check on the JS side with a dev time warning.

We don't have RenderScript path for blur filter yet, which I imagine might end up stealing some of this code later.

Changelog: [Internal]

Differential Revision: D55896447