Avoid Snapshotting Modal too often (#43995)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/43995

With a previous change, we started taking snapshots of the Modal every time a Fabric mounting event was happening. this might cause perf regression when a modal is presented, and there is no need to take a snapshot that often.

This change moves the snapshotting code right before the dismissal of the Modal.

## Changelog
[iOS][Changed] - Move the snapshotting code before the dismissal.

## Facebook
This should fix T179288585, T184520225

Differential Revision: D55914776

fbshipit-source-id: 6679babf7f72aef7254113497116d5482640e789