Disable view preallocation when batchRenderingUpdatesInEventLoop is enabled (#44144)
Summary:

When using `batchRenderingUpdatesInEventLoop` the Fabric Differentiator may more a newly created ShadowNode and subsequent update (eg coming from a layout effect) into a single MountingTransaction. This enables us to correctly implement layoutEffect in the new architecture and prevent flickering of intermediate states.

On Android, we also have a view preallocation mechanism, which will create native views for ShadowNodes as they're created by React. This mechanism is incompatible with `batchRenderingUpdatesInEventLoop` as we will ignore the `create` instruction in the MountingTransaction since we already have a preallocated view. However the props we used for preallocation will be outdated/different compared with the ones in the `MountingTransaction`, which represents a severe regression to the issue described in https://github.com/facebook/react-native/issues/44111.

Changelog: [Android][Fixed] Mutations to newly created views from useLayoutEffect were not always applied.

Differential Revision: D56301318