Make sure to return a color for API < 26 (#44224)
Summary:

When landing D55749058, I had to gate some API calls with a version check as those API where not available for Versions < 26.

However, while doing that, I overlooked the fact that we might end up not returning any color and throwing an exception.

This change makes sure that, if the version does not support API 26, we fall back to the previous mechanism.

## Changelog:
[Android][Fixed] - Make sure to return a color when APIs version do not support WideGamut

Reviewed By: arushikesarwani94

Differential Revision: D56485922