Allow custom override for setColorScheme and utilize UIScreen.mainScreen for default fallback
Summary:
This diff introduces two configurable APIs for the RCTAppearance module.

1. It allows a custom handler when setColorScheme() is called
2. It allows the default fallback style to be grabbed from UIScreen.mainScreen. Previously with the TraitCollections being passed in from overridden views it was not getting the accurate system fallback.

We need this for Twilight, which is adopting a Light/Dark mode toggle. Previously when setColorScheme was getting called it would modify overrideUserInterfaceStyle and that would serve as the "default fallback" for future setColorScheme calls. setColorScheme shouldn't be setting the defaults, it should be setting the user-session theme preference.

Changelog:
[Internal] [Changed] - Added option for custom Appearance setColorScheme() handler and treating the UIScreen.mainScreen's userInterfaceStyle as the source of truth for the system's dark/light mode.

Differential Revision: D56868862